refactor: removed hard_coded str, added email as an extra param
